test(routines): split service_trigger_tests.rs to clear the 500-line gate
PR #1014 lowered the pre-push line-count gate to 500 lines, which put
service_trigger_tests.rs (742 lines) over the limit. Split it into three
files by concern (trigger/scheduled-fire, snooze/lock/crontab-sync, and
ANSI log-stripping) with no behavior change.
